Modified to build libftdi with position indep code

+++ scripts/docker/build-libftdi-pkg.sh
@@ -22,18 +22,43 @@ export VERSION=1.4
 export URL="http://www.intra2net.com/en/developer/libftdi/download/libftdi1-1.4.tar.bz2"
 
 #Download it
-if [ ! -f "${NAME}_${VERSION}.tar.bz2" ]
+if [ ! -f "${NAME}-${VERSION}.tar.bz2" ]
 then
     echo "Downloading: $URL..."
     wget "$URL" 
 fi
 
-if [[ ! -d "${NAME}_${VERSION}" ]] ; then
+if [[ ! -d "${NAME}-${VERSION}" ]] ; then
     tar xjvf ${NAME}-${VERSION}.tar.bz2
 fi
 
-cd libftdi1-${VERSION}
-rm -rf $buildDir
-mkdir -p $buildDir && cd $buildDir
-cmake -DCMAKE_TOOLCHAIN_FILE=../crosstoolchain-$1.cmake -DCMAKE_INSTALL_PREFIX="/usr" ../
-make
+cd ${NAME}-${VERSION}
+
+# need to build this as place independant code, so add -fPIC flag to CMAKE_C_FLAGS
+grep -e "-fPIC" CMakeLists.txt
+#echo "grep results: $?"
+grepResults=$?
+if [[ "$grepResults" == 1 ]] ; then
+    echo "Didn't find -fPIC, adding it..."
+    sed -i 's/add_definitions(-Wall)/add_definitions(-Wall -fPIC)/' CMakeLists.txt
+elif [[ "$grepResults" == 0 ]] ; then
+    echo "-fPIC already in CMakeLists.txt. Carry on..."
+elif [[ "$grepResults" == 2 ]] ; then
+    echo "General grep error when searching CMakeLists.txt for -fPIC flag"
+    exit 100
+fi
+
+if [[ "$?" == 0 ]] ; then
+
+    # always a clean build...
+    rm -rf $buildDir
+    mkdir -p $buildDir && cd $buildDir
+    cmake -DCMAKE_TOOLCHAIN_FILE=/home/builder/libftdi/crosstoolchain-$CROSS_ARCH.cmake -DCMAKE_INSTALL_PREFIX="/usr" ../
+
+    # at this point we just need the static lib, not the package
+    make
+    #make package
+    #mv libftdi1-$VERSION.deb libftdi1-$VERSION-$CROSS_ARCH.deb
+else
+    echo "Failed to add -fPIC to CMakeLists.txt"
+fi
